wip share fields and transformers
do not share utils
