+++ bioimageio/spec/__init__.py
@@ -1,9 +1,4 @@
-from bioimageio.spec.common_utils import CommonVersionSpecificUtils
 from . import v0_1, v0_3
-from .latest import fields, maybe_convert_model, nodes, raw_nodes, schema, utils
+from .latest import *
 
-__version__ = nodes.FormatVersion.__args__[-1]
-
-load_and_resolve_spec = utils.load_and_resolve_spec
-load_model_spec = utils.load_model_spec
-load_spec = utils.load_spec
+__version__ = FormatVersion.__args__[-1]
