--- bioimageio/spec/common_utils/__init__.py
@@ -1,2 +1,0 @@
-from .common import yaml
-from ._common_version_specific import CommonVersionSpecificUtils
