+++ bioimageio/spec/latest/__init__.py
@@ -1,1 +1,2 @@
-from bioimageio.spec.v0_3 import fields, maybe_convert_model, nodes, raw_nodes, schema, utils  # noqa
+from bioimageio.spec.v0_3 import *  # noqa
+from .build_spec import build_spec
