+++ bioimageio/spec/latest/build_spec.py
@@ -20,7 +20,7 @@ def _get_local_path(uri, root=None):
             is_local_path = True
     if not is_local_path:
         uri = spec.fields.URI().deserialize(uri)
-        uri = spec.utils.download_uri_node_to_local_path(uri).as_posix()
+        uri = spec.download_uri_to_local_path(uri).as_posix()
     return uri
