+++ bioimageio/spec/shared/common.py
@@ -1,6 +1,11 @@
+import os
+import pathlib
+
 from ruamel.yaml import YAML
 
 yaml = YAML(typ="safe")
+
+BIOIMAGEIO_CACHE_PATH = pathlib.Path(os.getenv("BIOIMAGEIO_CACHE_PATH", pathlib.Path.home() / "bioimageio_cache"))
 
 
 class Singleton(type):
