+++ bioimageio/spec/v0_1/__init__.py
@@ -1,10 +1,13 @@
-from bioimageio.spec.common_utils import CommonVersionSpecificUtils
-from bioimageio.spec.v0_3 import fields  # noqa; not a mistake, v0_1 does not have its own fields
+from bioimageio.spec.shared import fields
 from . import nodes, raw_nodes, schema
 from .raw_nodes import FormatVersion
 
-utils = CommonVersionSpecificUtils(FormatVersion.__args__[-1])
+fields = fields
 
 
 def maybe_convert_model(data):
+    return data
+
+
+def maybe_convert_manifest(data):
     return data
