+++ bioimageio/spec/v0_3/__init__.py
@@ -1,6 +1,12 @@
-from bioimageio.spec.common_utils import CommonVersionSpecificUtils
-from . import fields, nodes, raw_nodes, schema
-from ._maybe_convert import maybe_convert_model
+from bioimageio.spec.shared import fields
+from . import nodes, raw_nodes, schema, utils
+from .converters import maybe_convert_manifest, maybe_convert_model
 from .raw_nodes import FormatVersion
 
-utils = CommonVersionSpecificUtils(FormatVersion.__args__[-1])
+fields = fields
+
+
+get_nn_instance = utils.get_nn_instance
+download_uri_to_local_path = utils.download_uri_to_local_path
+load_raw_model = utils.load_raw_model
+load_model = utils.load_model
