+++ bioimageio/spec/v0_3/raw_nodes.py
@@ -3,23 +3,30 @@ from __future__ import annotations
 import distutils.version
 from dataclasses import dataclass
 from datetime import datetime
-from pathlib import Path
-from typing import Any, Dict, List, NewType, Optional, TYPE_CHECKING, Tuple, Union
+from typing import Any, Dict, List, NewType, Optional, Tuple, Union
 
 from marshmallow import missing
+
+from bioimageio.spec.shared.raw_nodes import (
+    ImplicitInputShape,
+    ImplicitOutputShape,
+    ImportableModule,
+    ImportablePath,
+    Node,
+    URI,
+)
 
 try:
     from typing import Literal
 except ImportError:
     from typing_extensions import Literal
 
-if TYPE_CHECKING:
-    from . import schema
 
 # Ideally only the current format version is valid.
 # Older formats may be converter through `bioimageio.spec.utils.maybe_convert`,
 # such that we only need to support the most up-to-date version.
 FormatVersion = Literal["0.3.0", "0.3.1", "0.3.2"]  # newest format needs to be last (used in spec.__init__.py)
+latest_version = FormatVersion.__args__[-1]  # todo: py 3.9: use typing.get_args
 ManifestFormatVersion = Literal["0.1.0", "0.2.0"]  # newest format expected to be last
 
 PreprocessingName = Literal["binarize", "clip", "scale_linear", "sigmoid", "zero_mean_unit_variance", "scale_range"]
@@ -44,27 +51,10 @@ Axes = NewType("Axes", str)
 
 
 @dataclass
-class Node:
-    pass
-
-
-@dataclass
 class Author(Node):
     name: str = missing
     affiliation: Optional[str] = missing
     orcid: Optional[str] = missing
-
-
-@dataclass
-class ImportablePath(Node):
-    filepath: Path = missing
-    callable_name: str = missing
-
-
-@dataclass
-class ImportableModule(Node):
-    module_name: str = missing
-    callable_name: str = missing
 
 
 ImportableSource = Union[ImportableModule, ImportablePath]
@@ -75,27 +65,6 @@ class CiteEntry(Node):
     text: str = missing
     doi: Optional[str] = missing
     url: Optional[str] = missing
-
-
-@dataclass
-class URI(Node):
-    """URI as  scheme:[//authority]path[?query][#fragment]"""
-
-    scheme: str = missing
-    authority: str = missing
-    path: str = missing
-    query: str = missing
-    fragment: str = missing
-
-    def __str__(self):
-        """scheme:[//authority]path[?query][#fragment]"""
-        return (
-            (self.scheme + ":" if self.scheme else "")
-            + ("//" + self.authority if self.authority else "")
-            + self.path
-            + ("?" + self.query if self.query else "")
-            + ("#" + self.fragment if self.fragment else "")
-        )
 
 
 @dataclass
@@ -125,25 +94,6 @@ class RDF(Node):
     timestamp: datetime = missing
     type: Type = missing
     version: Optional[distutils.version.StrictVersion] = missing
-
-
-@dataclass
-class ImplicitInputShape(Node):
-    min: List[float] = missing
-    step: List[float] = missing
-
-    def __len__(self):
-        return len(self.min)
-
-
-@dataclass
-class ImplicitOutputShape(Node):
-    reference_input: str = missing
-    scale: List[float] = missing
-    offset: List[int] = missing
-
-    def __len__(self):
-        return len(self.scale)
 
 
 @dataclass
@@ -179,11 +129,6 @@ class OutputTensor:
     postprocessing: List[Postprocessing] = missing
     description: Optional[str] = missing
     data_range: Tuple[float, float] = missing
-
-
-@dataclass
-class SpecURI(URI):
-    spec_schema: schema.RDF = missing
 
 
 @dataclass
