+++ tests/conftest.py
@@ -3,7 +3,6 @@ from pathlib import Path
 import pytest
 from ruamel.yaml import YAML
 
-from bioimageio.spec import load_and_resolve_spec
 
 yaml = YAML(typ="safe")
 
@@ -14,16 +13,18 @@ def rf_config_path_v0_1():
 
 
 @pytest.fixture
-def rf_config_path_v0_3_0():
-    return Path(__file__).parent / "../specs/models/sklearn/RandomForestClassifier.model.yaml"
-
-
-rf_config_path = rf_config_path_v0_3_0
+def rf_config_path_v0_3_1():
+    return Path(__file__).parent / "../specs/models/sklearn/RandomForestClassifier_v0_3_1.model.yaml"
 
 
 @pytest.fixture
-def rf_resolved_spec(rf_config_path):
-    return load_and_resolve_spec(rf_config_path)
+def rf_config_path_v0_3_2():
+    return Path(__file__).parent / "../specs/models/sklearn/RandomForestClassifier_v0_3_2.model.yaml"
+
+
+@pytest.fixture
+def rf_config_path(rf_config_path_v0_3_2):
+    return rf_config_path_v0_3_2
 
 
 @pytest.fixture
