+++ tests/test_utils.py
@@ -3,10 +3,10 @@ from pathlib import Path
 from typing import Any
 
 import pytest
-from marshmallow import post_load
 from ruamel.yaml import YAML
 
-from bioimageio.spec import fields, load_model_spec, nodes, raw_nodes, schema, utils
+from bioimageio.spec import load_raw_model, nodes, raw_nodes
+from bioimageio.spec.shared import transformers
 
 yaml = YAML(typ="safe")
 
@@ -19,7 +19,7 @@ class MyNode(nodes.Node):
 
 def test_iter_fields():
     entry = MyNode("a", 42)
-    assert [("field_a", "a"), ("field_b", 42)] == list(utils._iter_fields(entry))
+    assert [("field_a", "a"), ("field_b", 42)] == list(transformers.iter_fields(entry))
 
 
 @dataclass
@@ -45,11 +45,11 @@ class TestNodeVisitor:
         )
 
     def test_node(self, tree):
-        visitor = utils._NodeVisitor()
+        visitor = transformers.NodeVisitor()
         visitor.visit(tree)
 
     def test_node_transform(self, tree):
-        class MyTransformer(utils._NodeTransformer):
+        class MyTransformer(transformers.NodeTransformer):
             def transform_URL(self, node):
                 return Content(f"content of url {node.url}")
 
@@ -59,44 +59,6 @@ class TestNodeVisitor:
         assert isinstance(transformed_tree.left.right, Content)
 
 
-@dataclass
-class MySpec(nodes.Node):
-    spec_uri_a: raw_nodes.SpecURI
-    spec_uri_b: raw_nodes.SpecURI
-
-
-class SubSpec(schema.PyBioSchema):
-    axes = fields.Axes()
-
-    @post_load
-    def make_object(self, data, **kwargs):
-        return data
-
-
-class Spec(schema.PyBioSchema):
-    spec_uri_a = fields.SpecURI(SubSpec)
-    spec_uri_b = fields.SpecURI(SubSpec)
-
-    @post_load
-    def make_object(self, data, **kwargs):
-        return MySpec(**data)
-
-
-class TestTraversingSpecURI:
-    def test_resolve_spec(self):
-        tree = Spec().load({"spec_uri_a": "https://example.com", "spec_uri_b": "../file.yml"})
-
-        class MyTransformer(utils._NodeTransformer):
-            def transform_SpecURI(self, node):
-                res = {"axes": "xyc"}
-                return node.spec_schema.load(res)
-
-        transformer = MyTransformer()
-        transformed_tree = transformer.transform(tree)
-        assert {"axes": "xyc"} == transformed_tree.spec_uri_a
-        assert {"axes": "xyc"} == transformed_tree.spec_uri_b
-
-
 def test_resolve_import_path(tmpdir):
     tmpdir = Path(tmpdir)
     manifest_path = tmpdir / "manifest.yaml"
@@ -104,8 +66,8 @@ def test_resolve_import_path(tmpdir):
     filepath = tmpdir / "my_mod.py"
     filepath.write_text("class Foo: pass", encoding="utf8")
     node = raw_nodes.ImportablePath(filepath=filepath, callable_name="Foo")
-    uri_transformed = utils._UriNodeTransformer(root_path=tmpdir).transform(node)
-    source_transformed = utils._SourceNodeTransformer().transform(uri_transformed)
+    uri_transformed = transformers.UriNodeTransformer(root_path=tmpdir).transform(node)
+    source_transformed = transformers.SourceNodeTransformer().transform(uri_transformed)
     assert isinstance(source_transformed, nodes.ImportedSource)
     Foo = source_transformed.factory
     assert Foo.__name__ == "Foo"
@@ -114,10 +76,10 @@ def test_resolve_import_path(tmpdir):
 
 def test_resolve_directory_uri(tmpdir):
     node = raw_nodes.URI(scheme="", authority="", path=str(tmpdir), query="", fragment="")
-    uri_transformed = utils._UriNodeTransformer(root_path=Path(tmpdir)).transform(node)
+    uri_transformed = transformers.UriNodeTransformer(root_path=Path(tmpdir)).transform(node)
     assert uri_transformed == Path(tmpdir)
 
 
-def test_load_model_spec(rf_config_path):
+def test_load_raw_model(rf_config_path):
     rf_model_data = yaml.load(rf_config_path)
-    load_model_spec(rf_model_data, rf_config_path)
+    load_raw_model(rf_model_data, rf_config_path)
